feat: 상품 삭제 API 컨트롤러에서 인증 확인

+++ src/main/java/hello/until/item/controller/ItemController.java
@@ -1,5 +1,8 @@
 package hello.until.item.controller;
 
+import hello.until.auth.PrincipalDetails;
+import hello.until.exception.CustomException;
+import hello.until.exception.ExceptionCode;
 import hello.until.item.dto.request.CreateItemRequest;
 import hello.until.item.dto.request.UpdateItemRequest;
 import hello.until.item.dto.response.ItemResponse;
@@ -8,6 +11,7 @@ import hello.until.item.service.ItemService;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -51,7 +55,12 @@ public class ItemController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteItem(@PathVariable Long id) {
+    public void deleteItem(
+            @AuthenticationPrincipal PrincipalDetails principalDetails,
+            @PathVariable Long id) {
+        if (principalDetails == null) {
+            throw new CustomException(ExceptionCode.NOT_AUTHENTICATED);
+        }
         this.itemService.deleteItem(id);
     }
 }
